requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.0
astroid==2.4.2
async-timeout==4.0.2
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
discord.py==2.1.0
Django==4.0.4
dnspython==2.1.0
frozenlist==1.3.3
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.7.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.7.1
mccabe==0.6.1
multidict==5.1.0
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pygame==2.1.2
pylint==2.6.0
pymongo==3.11.4
PyNaCl==1.5.0
pyOpenSSL==21.0.0
python-dotenv==0.17.1
queuelib==1.6.2
requests==2.25.1
Scrapy==2.5.1
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.7.4.3
tzdata==2022.1
urllib3==1.26.3
w3lib==1.22.0
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.12.17
zope.interface==5.4.0
